// src/gateway/routes/graphql-v2/schema/types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type RequireFields<T, K extends keyof T> = {
  [X in Exclude<keyof T, K>]?: T[X];
} & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type BlockHeightFilter = {
  /** Minimum block height to filter from */
  min?: Maybe<Scalars["Int"]>;
  /** Maximum block height to filter to */
  max?: Maybe<Scalars["Int"]>;
};
// ...
export type QueryTransactionsArgs = {
  ids?: Maybe<Array<Scalars["ID"]>>;
  owners?: Maybe<Array<Scalars["String"]>>;
  recipients?: Maybe<Array<Scalars["String"]>>;
  tags?: Maybe<Array<TagFilter>>;
  bundledIn?: Maybe<Array<Scalars["ID"]>>;
  block?: Maybe<BlockHeightFilter>;
  first?: Maybe<Scalars["Int"]>;
  after?: Maybe<Scalars["String"]>;
  sort?: Maybe<SortOrder>;
};

export type QueryBlockArgs = {
  id?: Maybe<Scalars["String"]>;
};

export type QueryBlocksArgs = {
  ids?: Maybe<Array<Scalars["ID"]>>;
  height?: Maybe<BlockHeightFilter>;
  first?: Maybe<Scalars["Int"]>;
  after?: Maybe<Scalars["String"]>;
  sort?: Maybe<SortOrder>;
};

/** Optionally reverse the result sort order from `HEIGHT_DESC` (default) to `HEIGHT_ASC`. */
export enum SortOrder {
  /**
   * Results are sorted by the transaction block height in ascending order, with
   * the oldest transactions appearing first, and the most recent and
   * pending/unconfirmed appearing last.
   */
  HeightAsc = "HEIGHT_ASC",
  /**
   * Results are sorted by the transaction block height in descending order, with
   * the most recent and unconfirmed/pending transactions appearing first.
   */
  HeightDesc = "HEIGHT_DESC",
}

export type Tag = {
  __typename?: "Tag";
  /** UTF-8 tag name */
  name: Scalars["String"];
  /** UTF-8 tag value */
  value: Scalars["String"];
};

/** Find transactions with the folowing tag name and value */
export type TagFilter = {
  /** The tag name */
  name: Scalars["String"];
  /**
   * An array of values to match against. If multiple values are passed then
   * transactions with _any_ matching tag value from the set will be returned.
   *
   * e.g.
   *
   * \`{name: "app-name", values: ["app-1"]}\`
   *
   * Returns all transactions where the \`app-name\` tag has a value of \`app-1\`.
   *
   * \`{name: "app-name", values: ["app-1", "app-2", "app-3"]}\`
   *
   * Returns all transactions where the \`app-name\` tag has a value of either \`app-1\` _or_ \`app-2\` _or_ \`app-3\`.
   */
  values: Array<Scalars["String"]>;
  /** The operator to apply to to the tag filter. Defaults to EQ (equal). */
  op?: Maybe<TagOperator>;
};

/** The operator to apply to a tag value. */
export enum TagOperator {
  /** Equal */
  Eq = "EQ",
  /** Not equal */
  Neq = "NEQ",
}